user permission bug fixed

+++ app/Http/Controllers/CategoryController.php
@@ -24,7 +24,7 @@ class CategoryController extends Controller
    // add Category view page
    public function AddCategory()
    {
-    if(is_null($this->user) || !$this->user->can('product.create') || !$this->user->can('product.update') || !$this->user->can('product.delete') || !$this->user->can('product.view')){
+    if(is_null($this->user)|| !$this->user->can('product.delete')){
         abort('403','You dont have acces!!!!');
     }
 
@@ -36,6 +36,9 @@ class CategoryController extends Controller
 
         public function StoreCategory(Request $request)
         {
+            if(is_null($this->user)|| !$this->user->can('product.delete')){
+                abort('403','You dont have acces!!!!');
+            }
             $validateData = $request->validate([
                 'category_name' => 'required',
 
@@ -59,18 +62,27 @@ class CategoryController extends Controller
 
    public function CategoryList()
    {
+    if(is_null($this->user)|| !$this->user->can('product.delete')){
+        abort('403','You dont have acces!!!!');
+    }
     $categories  = Category::all();
        return view('category.CategoryList',compact('categories'));
    }
 
    public function EditCategory($id)
 {
+    if(is_null($this->user)|| !$this->user->can('product.delete')){
+        abort('403','You dont have acces!!!!');
+    }
     $category = Category::find($id);
     return view('category.CategoryEdit',compact('category'));
 }
 
 public function UpdateCategory(Request $request,$id)
 {
+    if(is_null($this->user)|| !$this->user->can('product.delete')){
+        abort('403','You dont have acces!!!!');
+    }
     $validateData = $request->validate([
     'category_name' => 'required',
 
@@ -87,6 +99,9 @@ public function UpdateCategory(Request $request,$id)
 }
 public function DeleteCategory($id)
 {
+    if(is_null($this->user)|| !$this->user->can('product.delete')){
+        abort('403','You dont have acces!!!!');
+    }
     Category::destroy($id);
     return redirect('category/list');
